refactor: remove duplicated styles
