+++ src/app/my-page/panel/NavBar.style.ts
@@ -1,5 +1,3 @@
-import { SxProps } from '@mui/material'
-
 export const whale = {
   fontSize: '1.0625rem', // = width?
 }
@@ -18,15 +16,4 @@ export const silhouette = {
 
 export const setting = {
   fontsize: '0.75rem',
-}
-
-export const pcNavBar: SxProps = {
-  width: '30%',
-  maxWidth: '19.25rem',
-  height: 'content-fit',
-}
-
-export const mobileNavBar: SxProps = {
-  width: '100%',
-  height: 'content-fit',
 }
