+++ src/components/NavBarBox.style.ts
@@ -1,0 +1,12 @@
+import { SxProps } from '@mui/material'
+
+export const pcNavBar: SxProps = {
+  width: '30%',
+  maxWidth: '19.25rem',
+  height: 'content-fit',
+}
+
+export const mobileNavBar: SxProps = {
+  width: '100%',
+  height: 'content-fit',
+}
